use `oxnet::{ IpNet, Ipv4Net, Ipv6Net }` (#5810)

+++ nexus/test-utils/src/lib.rs
@@ -675,7 +675,7 @@ impl<'a, N: NexusServer> ControlPlaneTestContextBuilder<'a, N> {
                 nic: NetworkInterface {
                     id: Uuid::new_v4(),
                     ip: NEXUS_OPTE_IPV4_SUBNET
-                        .nth(NUM_INITIAL_RESERVED_IP_ADDRESSES as u32 + 1)
+                        .nth(NUM_INITIAL_RESERVED_IP_ADDRESSES + 1)
                         .unwrap()
                         .into(),
                     kind: NetworkInterfaceKind::Service {
@@ -1029,7 +1029,7 @@ impl<'a, N: NexusServer> ControlPlaneTestContextBuilder<'a, N> {
                     nic: NetworkInterface {
                         id: Uuid::new_v4(),
                         ip: DNS_OPTE_IPV4_SUBNET
-                            .nth(NUM_INITIAL_RESERVED_IP_ADDRESSES as u32 + 1)
+                            .nth(NUM_INITIAL_RESERVED_IP_ADDRESSES + 1)
                             .unwrap()
                             .into(),
                         kind: NetworkInterfaceKind::Service {
